Moved OpenAiApi to its own package.
When consuming this library from a Java 11 modular application, no
classes in the same package must be exported from multiple JARs.
