+++ client/src/main/java/com/theokanning/openai/AuthenticationInterceptor.java
@@ -1,28 +1,15 @@
 package com.theokanning.openai;
 
-import okhttp3.Interceptor;
-import okhttp3.Request;
-import okhttp3.Response;
-
-import java.io.IOException;
-
 /**
  * OkHttp Interceptor that adds an authorization token header
+ * 
+ * @deprecated Use {@link com.theokanning.openai.client.AuthenticationInterceptor}
  */
-public class AuthenticationInterceptor implements Interceptor {
-
-    private final String token;
+@Deprecated
+public class AuthenticationInterceptor extends com.theokanning.openai.client.AuthenticationInterceptor {
 
     AuthenticationInterceptor(String token) {
-        this.token = token;
+        super(token);
     }
 
-    @Override
-    public Response intercept(Chain chain) throws IOException {
-        Request request = chain.request()
-                .newBuilder()
-                .header("Authorization", "Bearer " + token)
-                .build();
-        return chain.proceed(request);
-    }
 }
