+++ client/src/main/java/com/theokanning/openai/client/AuthenticationInterceptor.java
@@ -1,0 +1,28 @@
+package com.theokanning.openai.client;
+
+import java.io.IOException;
+
+import okhttp3.Interceptor;
+import okhttp3.Request;
+import okhttp3.Response;
+
+/**
+ * OkHttp Interceptor that adds an authorization token header
+ */
+public class AuthenticationInterceptor implements Interceptor {
+
+    private final String token;
+
+    protected AuthenticationInterceptor(String token) {
+        this.token = token;
+    }
+
+    @Override
+    public Response intercept(Chain chain) throws IOException {
+        Request request = chain.request()
+                .newBuilder()
+                .header("Authorization", "Bearer " + token)
+                .build();
+        return chain.proceed(request);
+    }
+}
