+++ client/src/test/java/com/theokanning/openai/CompletionTest.java
@@ -1,5 +1,6 @@
 package com.theokanning.openai;
 
+import com.theokanning.openai.client.OpenAiService;
 import com.theokanning.openai.completion.CompletionChoice;
 import com.theokanning.openai.completion.CompletionRequest;
 import org.junit.jupiter.api.Test;
