+++ client/src/test/java/com/theokanning/openai/ModerationTest.java
@@ -1,6 +1,7 @@
 package com.theokanning.openai;
 
 import com.theokanning.openai.moderation.ModerationRequest;
+import com.theokanning.openai.client.OpenAiService;
 import com.theokanning.openai.moderation.Moderation;
 import org.junit.jupiter.api.Test;
